requirements.txt
requests==2.31.0
pymongo==4.6.1
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0
streamlit==1.29.0
plotly==5.18.0
shap==0.44.0
schedule==1.2.0
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
